Euler.Problems

Documentation

answers :: Map Integer (IO String)

Answers to Euler problems.

The answers to Project Euler problems are all numeric, but the type here is String rather than Integer because in some cases the answer is semantically a "list of digits" rather than an integer, which may be an important distinction if the first digit of an answer is zero.

Most of the answers are pure, but IO is allowed because some problems involve reading input data from files.