Euler.Problems.Problem19

Documentation

data Date

Constructors

Date

Fields

dateYear :: Int

Anno Domini

dateMonth :: Int

1 (january) to 12 (december)

dateDay :: Int

1 to 31

dateWeekday :: Int

1 (money) to 7 (sunday)

datesFrom1900 :: [Date]

All dates starting from Jan 1, 1900.

answer :: Integer

The number of Sundays that fell on the first of the month during the twentieth century (1 Jan 1901 to 31 Dec 2000).