Euler.Problems.Problem23

Description

A number n is called abundant if the sum of its proper divisors exceeds n.

Documentation

answer :: Integer

The sum of all the positive integers which cannot be written as the sum of two abundant numbers.

answerBounded

Arguments

:: Integer
bound

-> Integer

The sum of all the positive integers ≤ bound which cannot be written as the sum of two abundant numbers.