Euler.Problems.Problem31

Description

In England the currency is made up of pound, £, and pence, p. There are eight coins in general circulation: 1p, 2p, 5p, 10p, 20p, 50p, £1 (100p), and £2 (200p).

Documentation

answer :: Natural

The number of ways to make £2 using any number of coins.

ways

Arguments

:: Natural
Target amount

-> [Natural]
Denominations (in descending order for best performance)

-> Natural
Number of ways to make the target amount using the denominations