Euler.Util.Date

Description

Dates are used in Euler problem 19.

Functions

monthLength :: Int -> Int -> Int

monthLength y m is the number of days in month m of year y.

monthLengths :: Int -> [Int]

yearLength :: Int -> Int

The number of days in a year AD.

leap :: Int -> Bool

Whether a year AD is a leap year. Leap years contain an extra day in February.

Properties

The length of a year is equal to the sum of the lengths of its months.

yearLength y == sum (monthLengths y)