Euler.Util.Inf

Documentation

data Inf a

A wrapper for infinite lists, with typeclasses operating on only the list's head. This is not correct in all circumstances, but is useful for (for example) tails of an infinite sequence without duplicate elements.

fromList :: [a] -> Inf a

toList :: Inf a -> [a]